requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.1.0
jinja2>=3.1.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0